{feature} support passing in the path to a builder directly
This makes it a lot easier to write non-reusable builders to
handle certain tasks in a repository. Rather than having to
pass in the path to a package that references a builder JSON
file that then references a builder, you can just point to
the builder.

This change also makes it possible to point to a builder JSON
file directly instead of a package, in case you have an entire
collection of non-reusable builders in your workspace.