{fix} stop trying to support "extra arguments"
Supporting this came down to magic, especially with
the way the Proxy works in clipanion.
That isn't really meant for catching unknown
arguments, it's for proxying arguments to an underlying
command.

To give an example, suppose there's a command with
an "arg" positional argument, a "name" string option
and a proxy.
If the command is called with
  --name "A name" "value for arg"
the value of the name option will be "A name", but if
the command is called with
  "value for arg" --name "A name"
the name option will be empty and ["--name", "A name"]
is passed to the proxy.