{refactor} rename atelier -> snuggery
Atelier is, sadly, a common name. Snuggery, meanwhile,
is a word that has fallen in relative disuse.

The connotation is different too. While an atelier is
the workspace of a master artist or craftsperson, a
snuggery is "a cozy room". A snuggery is a place where
you'd go and feel at home, where you\d feel safe.