{feature} create own configuration types
These types allow adding extensions to the TargetDefinition,
which makes it possible to use the types to fully manipulate
`workspace.json` files with `dependsOn` and `outputs` on the
target objects.

This commit still permits writing @angular-devkit/core types
and it provides a compatibility layer in case a workspace
object has to be passed into an API that expects a workspace
from @angular-devkit/core.