{feature} rewrite KDL support
Instead of tackling this using proxies to use
the underlying KDL Document as data, transform
the KDL document to a JSON object which is then
passsed to the existing Angular workspace definition.

The main reason to move away from proxies here
is that logging a proxy shows the unproxied value,
which is very confusing when debugging.

Secondly, transforming from KDL to JSON or back is
relatively easy. The difficulty lies in applying
changes to the KDL document based on changes on the
JSON objects. With proxies, the code for "easy"
transforms and the code for applying changes was
heavily mixed in the same functions.
The current approach has fromJson/toJson functions
perform the easy part, and add/modify/delete
functions for the more complex application of
changes.

Switch out the (official) kdljs parser with
a parser that preserves whitespace and comments.