{feature} don't parse workspace configuration to find location
Parsing the configuration there means we're at least doing
so twice, and one of the two results is entirely ignored.

Secondly, doing this using the globally installed @snuggery/snuggery
means it's impossible to use newer workspace configuration formats.

By just looking for the known workspace files without parsing them,
we limit the amount of changes in the configuration that break
global installations to "only break when file names are added/removed".
Limiting the number of breaking changes is important, because there
can only be one @snuggery/global, and it might have to serve many
different local @snuggery/snuggery versions.