{feature} rewrite custom JSON in KDL handling to closely match JiK 3
The new "JSON-in-KDL" (JiK) specification already matches the previous
implementation pretty well, except for arrays. The behaviour of defining
multiple nodes with the same name turns them into arrays is not present
in JiK, and frankly I wasn't really happy with it because it lead to
some very complex code to keep track of all the array items. It was also
confusing because

```kdl
someProp 1 2 3
```

is an array `[1, 2, 3]` while

```kdl
someProp 1 2 3
someProp 4
```

became the nested array `[[1, 2, 3], 4]`.

With that behaviour gone, managing changes in the JSON part has become
more straightforward.

Note we still deviate from JiK in one spot: we allow arrays to have one
child node not called "-": the "super" element in extended arrays.
Other than that the JSON handling perfectly matches the JiK specification.