{feature} build angular packages in a single pass
Remove the setup with multiple package.json files, and only
support multiple entry points via the package's exports property.

The new compilation doesn't use the `flatModuleOutFile` option
that Angular prescribes, because it limits us to building a single
entry file.
It doesn't look like disabling this mode removes much from the
angular compilation. We do now lack the validation angular adds
to ensure that all exposed directives/components are exported,
but with modules in decline the ergonomy of this new compilation
mode makes it worth it imo.

And in the end that's all that matters, as this is just for me :)