{feature} build-angular: add bundleDependencies option to library builder
This adds a new 'bundleDependencies' option to the library builder.
When enabled, it allows esbuild to bundle dependencies instead of marking them as external.
This is required for scenarios like running Karma tests with coverage, where dependencies need to be included in the bundle for the browser to resolve them correctly.